moved individual configurations to separate py files and reorganized core file structure
