+++ pleiades/configurations/__init__.py
@@ -1,0 +1,3 @@
+from pleiades.configurations.brb import *
+from pleiades.configurations.pcx import *
+from pleiades.configurations.wham import *
